refactor: move command into operator sub-commands
