+++ pkg/operator/split/flags.go
@@ -1,0 +1,20 @@
+package split
+
+import "github.com/urfave/cli/v2"
+
+var (
+	OperatorSplitFlag = cli.IntFlag{
+		Name:     "operator-split",
+		Aliases:  []string{"os"},
+		Usage:    "Split for the operator in bips (e.g. 1000 = 10%)",
+		Required: false,
+		EnvVars:  []string{"OPERATOR_SPLIT"},
+	}
+
+	AVSAddressFlag = cli.StringFlag{
+		Name:    "avs-address",
+		Aliases: []string{"aa"},
+		Usage:   "AVS address to set operator split",
+		EnvVars: []string{"AVS_ADDRESS"},
+	}
+)
