+++ pkg/rewards.go
@@ -14,8 +14,6 @@ func RewardsCmd(p utils.Prompter) *cli.Command {
 			rewards.ClaimCmd(p),
 			rewards.SetClaimerCmd(p),
 			rewards.ShowCmd(p),
-			rewards.SetOperatorSplitCmd(p),
-			rewards.GetOperatorSplitCmd(p),
 		},
 	}
